Pull spark configuration out of SparkOnHit component
