+++ Content.Shared/_ES/Sparks/ESSparksSystem.cs
@@ -1,6 +1,8 @@
 using Content.Shared._ES.Sparks.Components;
 using Content.Shared._ES.TileFires;
 using Content.Shared.Physics;
+using Content.Shared.Power.Components;
+using Content.Shared.Power.EntitySystems;
 using Content.Shared.Throwing;
 using JetBrains.Annotations;
 using Robust.Shared.Map;
@@ -11,16 +13,45 @@ using Robust.Shared.Timing;
 
 namespace Content.Shared._ES.Sparks;
 
-public sealed class ESSparksSystem : EntitySystem
+public sealed partial class ESSparksSystem : EntitySystem
 {
     [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly INetManager _net = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
+    [Dependency] private readonly SharedPowerReceiverSystem _powerReceiver = default!;
     [Dependency] private readonly ESSharedTileFireSystem _tileFire = default!;
     [Dependency] private readonly ThrowingSystem _throwing = default!;
     [Dependency] private readonly SharedTransformSystem _transform = default!;
 
     public static readonly EntProtoId DefaultSparks = "ESEffectSparks";
+
+    /// <summary>
+    /// Variant of <see cref="DoSparks(EntityUid, int, Nullable{EntProtoId}, Nullable{EntityUid}, float, bool)"/> that takes
+    /// the configuration from a base component rather than being passed in as args
+    /// </summary>
+    /// <param name="ent">Entity that the sparks are originating from. Additionally, holds YAML configuration for spark effect</param>
+    /// <param name="user">A "user" who triggered the sparks</param>
+    /// <param name="cooldown">If true, will check the cooldown on <see cref="ESSparkCooldownComponent"/> before spawning sparks</param>
+    public void DoSparks<T>(Entity<T?> ent, EntityUid? user = null, bool cooldown = true) where T : ESBaseSparkConfigurationComponent
+    {
+        if (!Resolve(ent, ref ent.Comp))
+            return;
+
+        if (!_random.Prob(ent.Comp.Prob))
+            return;
+
+        SharedApcPowerReceiverComponent? powerReceiver = null;
+        if (_powerReceiver.ResolveApc(ent, ref powerReceiver) &&
+            (!_powerReceiver.IsPowered((ent, powerReceiver)) || powerReceiver.Load <= 0))
+            return;
+
+        DoSparks(ent,
+            number: ent.Comp.Count,
+            ent.Comp.SparkPrototype,
+            user: user,
+            tileFireChance: ent.Comp.TileFireChance,
+            cooldown: cooldown);
+    }
 
     /// <summary>
     /// Spawns sparks originating from a target entity
